+++ .gitignore
@@ -8,4 +8,4 @@ __pycache__
 /build/
 /local/
 /dist/
-mappercore/static
+mapper_interactive/static
