+++ README.md
@@ -1,25 +1,26 @@
-# MapperCore
+# Mapper Interactive
 
-[![PyPI version](https://badge.fury.io/py/mappercore.svg)](https://badge.fury.io/py/mappercore)
-[![Build Status](https://travis-ci.com/mappercore/mappercore.svg?branch=master)](https://travis-ci.com/mappercore/mappercore)
+[![PyPI version](https://badge.fury.io/py/mapper_interactive.svg)](https://badge.fury.io/py/mapper_interactive)
+[![Build Status](https://travis-ci.org/MapperInteractive/MapperInteractive.svg?branch=master)](https://travis-ci.com/MapperInteractive/MapperInteractive)
 
-The Mapper algorithm is a method for topological data analysis invented by Gurjeet Singh, Facundo Mémoli and Gunnar Carlsson [^1]. MapperCore is a web-based visualization of Mapper algorithm. It is open source software and is released under the MIT License.
+The Mapper algorithm is a method for topological data analysis invented by Gurjeet Singh, Facundo Mémoli and Gunnar Carlsson [^1]. 
+Mapper Interactive is a web-based visualization of Mapper algorithm. It is open source software and is released under the MIT License.
 
 ## Live Demo
 
-[Here](https://github.com/yaodong/mapper-example) is an example project and a [live demo](https://mapper-example.herokuapp.com).
+[Here](https://github.com/MapperInteractive/live-example/) is an example project and a [live demo](https://mapper-example.herokuapp.com).
 
 ## Installation
 
-MapperCore can be installed from PyPI by:
+Mapper Interactive can be installed from PyPI by:
 
 ```shell
-pip install mappercore==1.2.6
+pip install mapper_interactive
 ```
 
 ## Documentation
 
-Please checkout our [wiki](https://github.com/yaodong/mapper-core/wiki) or [our website](https://mappercore.org) for documentaion.
+Please checkout our [website](http://mapper-interactive.org) for documentation.
 
 ## License
 
@@ -30,6 +31,5 @@ This project is licensed under the MIT License - see the `LICENSE` file for deta
 Pull requests are welcomed. 
 
 
-
-[^1]: Gurjeet Singh, Facundo Mémoli and Gunnar Carlsson, *Topological Methods for the Analysis of High Dimensional Data Sets and 3D Object Recognition*, Eurographics Symposium on Point Based Graphics, European Association for Computer Graphics, 2007, pp. 91–100, DOI [10.2312/SPBG/SPBG07/091-100](http://dx.doi.org/10.2312/SPBG/SPBG07/091-100)
+> [^1]: Gurjeet Singh, Facundo Mémoli and Gunnar Carlsson, *Topological Methods for the Analysis of High Dimensional Data Sets and 3D Object Recognition*, Eurographics Symposium on Point Based Graphics, European Association for Computer Graphics, 2007, pp. 91–100, DOI [10.2312/SPBG/SPBG07/091-100](http://dx.doi.org/10.2312/SPBG/SPBG07/091-100)
