+++ docs/index.rst
@@ -1,17 +1,18 @@
-.. MapperCore documentation master file, created by
+.. MapperInteractive documentation master file, created by
    sphinx-quickstart on Thu Mar 14 20:59:27 2019.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
 |PyPI version| |Build Status| |License: MIT|
 
-MapperCore
--------------
+Mapper Interactive
+------------------
 
 
-The Mapper algorithm is a method for topological data analysis invented by Gurjeet Singh, Facundo Mémoli and Gunnar Carlsson [^1]. MapperCore is a web-based visualization of Mapper algorithm. It is open source software and is released under the MIT License.
+The Mapper algorithm is a method for topological data analysis invented by Gurjeet Singh, Facundo Mémoli and Gunnar Carlsson [^1].
+Mapper Interactive is a web-based visualization of Mapper algorithm. It is open source software and is released under the MIT License.
 
-Please find the source `on Github <https://github.com/mappercore/mappercore>`_.
+Please find the source `on Github <https://github.com/MapperInteractive/MapperInteractive>`_.
 
 We provide out of the box configuration for using Kepler Mapper on the backend. 
 
@@ -19,11 +20,11 @@ We provide out of the box configuration for using Kepler Mapper on the backend.
 Installation
 ==============
 
-MapperCore can be installed from PyPI by:
+Mapper Interactive can be installed from PyPI by:
 
 ::
 
-    pip install mappercore
+    pip install mapper_interactive
 
 
 
@@ -44,11 +45,11 @@ MapperCore can be installed from PyPI by:
     km/index
 
 
-.. |PyPI version| image:: https://badge.fury.io/py/mappercore.svg
-   :target: https://badge.fury.io/py/mappercore
+.. |PyPI version| image:: https://badge.fury.io/py/mapper_interactive.svg
+   :target: https://badge.fury.io/py/mapper_interactive
 
-.. |Build Status| image:: https://travis-ci.com/mappercore/mappercore.svg?branch=master
-   :target: https://travis-ci.com/mappercore/mappercore
+.. |Build Status| image:: https://travis-ci.org/MapperInteractive/MapperInteractive.svg?branch=master
+   :target: https://travis-ci.org/MapperInteractive/MapperInteractive
 
 .. |License: MIT| image:: https://img.shields.io/badge/License-MIT-yellow.svg
    :target: https://opensource.org/licenses/MIT)
