+++ docs/km/index.rst
@@ -3,7 +3,7 @@ Kepler Mapper Interactive
 
 We provide off-the-shelf configuration for using Kepler Mapper as the Mapper construction engine. The only customization required is to load the data and pass it to the `KeplerMapperConf` object. 
 
-Following the :ref:`Getting Started Guide` to initialize the basic setup, see the :ref:`KM Configuration` page for more advanced customization, or see :ref:`Configurations` page to learn how to extend the MapperCore framework for yourself.
+Following the :ref:`Getting Started Guide` to initialize the basic setup, see the :ref:`KM Configuration` page for more advanced customization, or see :ref:`Configurations` page to learn how to extend the Mapper Interactive framework for yourself.
 
 
 .. toctree::
