+++ mapper_interactive/conf/kepler_mapper.py
