+++ static/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mappercore",
+  "name": "mapper_interactive",
   "version": "0.1.0",
   "description": "",
   "main": "index.js",
